problem2_rag_eval/requirements.txt
langchain
langchain-community
google-generativeai
chromadb
python-dotenv
sentence-transformers
supabase
fastapi
uvicorn[standard]
streamlit
requests
black
flake8
unstructured